SAGE-1619: re-enable GPU pod access with K3S v1.25.4
Without this environment variable definition only the 'l4t.csv' file listed
mounts are performed when using the NVidia runtime when launching pods.
This causes for many core (i.e. including /usr/local/cuda/) mounts to not
exist in the launched pod causing for GPU access to fail. This change
informs the runtime to perform the mounts form all '*.csv' files (as was
done with v1.20.2 K3S).